Setting module in progress.

+++ source/setting/src/main/java/org/archboy/clobaframe/setting/system/impl/JsonSystemSettingProvider.java
@@ -7,37 +7,43 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.util.HashMap;
+import java.util.LinkedHashMap;
 import java.util.Map;
 import org.apache.commons.io.IOUtils;
-import org.archboy.clobaframe.setting.impl.AbstractJsonSettingRepository;
+import org.archboy.clobaframe.setting.impl.AbstractJsonSettingAccess;
 import org.archboy.clobaframe.setting.impl.Support;
 import org.archboy.clobaframe.setting.system.SystemSettingProvider;
+import org.archboy.clobaframe.setting.system.SystemSettingRepository;
 
 /**
  *
  * @author yang
  */
-public class CustomSystemSettingProvider extends AbstractJsonSettingRepository implements SystemSettingProvider {
+public class JsonSystemSettingProvider extends AbstractJsonSettingAccess implements SystemSettingProvider, SystemSettingRepository {
 
 	private String dataDir;
 	private String fileName;
-	private Map<String, Object> setting = new HashMap<String, Object>();
 	
-	public CustomSystemSettingProvider(String dataDir, String fileName) {
+	public JsonSystemSettingProvider(String dataDir, String fileName) {
 		this.dataDir = dataDir;
 		this.fileName = fileName;
-		load();
 	}
 
-	private void load(){
+	@Override
+	public int getPriority() {
+		return PRIORITY_HIGH;
+	}
+
+	@Override
+	public Map<String, Object> getAll() {
 		File folder = new File(dataDir);
 		if (folder.exists()) {
 			File file = new File(folder, fileName);
-			if (file.exists()) {
+			if (file.exists() && file.isFile()) {
 				InputStream in = null;
 				try{
 					in = new FileInputStream(file);
-					setting = read(in);
+					return read(in);
 				}catch(IOException e){
 					// ignore
 				}finally{
@@ -45,27 +51,14 @@ public class CustomSystemSettingProvider extends AbstractJsonSettingRepository i
 				}
 			}
 		}
-	}
-	
-	@Override
-	public int getPriority() {
-		return PRIORITY_HIGH;
+		
+		return new LinkedHashMap<String, Object>();
 	}
 
 	@Override
-	public Map<String, Object> get() {
-		return setting;
-	}
-
-	@Override
-	public boolean canWrite() {
-		return true;
-	}
-
-	@Override
-	public void set(Map<String, Object> item) {
-		Support.merge(setting, item);
-
+	public void update(Map<String, Object> item) {
+		Map<String, Object> map = Support.merge(getAll(), item);
+		
 		// save
 		OutputStream out = null;
 		try {
@@ -75,8 +68,11 @@ public class CustomSystemSettingProvider extends AbstractJsonSettingRepository i
 			}
 
 			File file = new File(folder, fileName);
-			out = new FileOutputStream(file);
-			write(out, setting);
+			if (!file.exists() || 
+				(file.exists() && file.isFile())){
+				out = new FileOutputStream(file);
+				write(out, item);
+			}
 		} catch (IOException e) {
 			// ignore
 		} finally {
@@ -85,10 +81,10 @@ public class CustomSystemSettingProvider extends AbstractJsonSettingRepository i
 	}
 
 	@Override
-	public void set(String key, Object value) {
-		Map<String, Object> map = new HashMap<String, Object>();
+	public void update(String key, Object value) {
+		Map<String, Object> map = new LinkedHashMap<String, Object>();
 		map.put(key, value);
-		set(map);
+		update(map);
 	}
 	
 }
